kafka-demo/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
filelock==3.7.1
idna==3.3
kafka-python==2.0.2
lxml==4.9.0
requests==2.28.0
urllib3==1.26.9
